Refactor editor manager and dungeon editor for improved clarity and performance
- Removed unused functions and streamlined shortcut registration in EditorManager for better readability.
- Updated lambda captures to use 'this' pointer for consistency and clarity.
- Refactored DungeonEditor to utilize ranges for sorting and filling operations, enhancing performance and readability.
- Simplified table setup in DungeonEditor by using a static array for tool names, improving maintainability.

+++ src/app/gui/color.cc
@@ -8,12 +8,8 @@ namespace yaze {
 namespace gui {
 
 ImVec4 ConvertSnesColorToImVec4(const gfx::SnesColor& color) {
-  return ImVec4(static_cast<float>(color.rgb().x) / 255.0f,
-                static_cast<float>(color.rgb().y) / 255.0f,
-                static_cast<float>(color.rgb().z) / 255.0f,
-                1.0f  // Assuming alpha is always fully opaque for SNES colors,
-                      // adjust if necessary
-  );
+  return ImVec4(color.rgb().x / 255.0f, color.rgb().y / 255.0f,
+                color.rgb().z / 255.0f, 1.0f);
 }
 
 gfx::SnesColor ConvertImVec4ToSnesColor(const ImVec4& color) {
